Move data classes and utils to core folder
